Demo App restructured (#2059)
* DemoApp is now restructured
