+++ DemoApp/AppDelegate.swift
@@ -2,23 +2,22 @@
 // Copyright © 2022 Stream.io Inc. All rights reserved.
 //
 
-import Sentry
 import StreamChat
 import UIKit
 
 @main
 class AppDelegate: UIResponder, UIApplicationDelegate {
+    var chat: StreamChatWrapper {
+        StreamChatWrapper.shared
+    }
+
     func application(
         _ application: UIApplication,
         didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?
     ) -> Bool {
-        #if RELEASE
-        // We're tracking Crash Reports / Issues from the Demo App to keep improving the SDK
-        SentrySDK.start { options in
-            options.dsn = "https://75b1074a38704dc0923d9de56fe1e1d4@o389650.ingest.sentry.io/6379288"
-            options.tracesSampleRate = 1.0
-        }
-        #endif
+        // Sentry
+        DemoAppConfiguration.configureSentry()
+
         return true
     }
 
@@ -26,11 +25,8 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         _ application: UIApplication,
         didRegisterForRemoteNotificationsWithDeviceToken deviceToken: Data
     ) {
-        ChatClient.shared.currentUserController().addDevice(.apn(token: deviceToken)) {
-            if let error = $0 {
-                log.error("adding a device failed with an error \(error)")
-            }
-        }
+        // Registers current device for push notifications on Stream backend
+        chat.registerForPushNotifications(with: deviceToken)
     }
 
     // MARK: UISceneSession Lifecycle
