+++ DemoApp/DemoAppCoordinator.swift
@@ -2,29 +2,27 @@
 // Copyright © 2022 Stream.io Inc. All rights reserved.
 //
 
-import Atlantis
-import GDPerformanceView_Swift
 import StreamChat
 import StreamChatUI
 import UIKit
 
-extension ChatClient {
-    static var shared: ChatClient!
-}
-
-private var isStreamInternalConfiguration: Bool {
-    ProcessInfo.processInfo.environment["STREAM_DEV"] != nil
-}
-
 final class DemoAppCoordinator: NSObject {
-    let window: UIWindow
+    private let window: UIWindow
+    private let chat: StreamChatWrapper
+    private let pushNotifications: PushNotifications
 
-    init(window: UIWindow) {
+    init(
+        window: UIWindow,
+        chat: StreamChatWrapper,
+        pushNotifications: PushNotifications
+    ) {
         self.window = window
+        self.chat = chat
+        self.pushNotifications = pushNotifications
         
         super.init()
-        
-        UNUserNotificationCenter.current().delegate = self
+
+        handlePushNotificationResponse()
     }
     
     func start(cid: ChannelId? = nil) {
@@ -34,33 +32,20 @@ final class DemoAppCoordinator: NSObject {
             showLogin(animated: false)
         }
     }
-}
 
-// MARK: - UNUserNotificationCenterDelegate
+    func handlePushNotificationResponse() {
+        pushNotifications.onNotificationResponse = { [weak self] response in
+            guard case UNNotificationDefaultActionIdentifier = response.actionIdentifier else {
+                return
+            }
+            guard
+                let chatNotificationInfo = self?.chat.notificationInfo(for: response),
+                let cid = chatNotificationInfo.cid else {
+                return
+            }
 
-extension DemoAppCoordinator: UNUserNotificationCenterDelegate {
-    func userNotificationCenter(
-        _ center: UNUserNotificationCenter,
-        didReceive response: UNNotificationResponse,
-        withCompletionHandler completionHandler: @escaping () -> Void
-    ) {
-        defer {
-            completionHandler()
+            self?.start(cid: cid)
         }
-
-        guard let notificationInfo = try? ChatPushNotificationInfo(content: response.notification.request.content) else {
-            return
-        }
-
-        guard let cid = notificationInfo.cid else {
-            return
-        }
-
-        guard case UNNotificationDefaultActionIdentifier = response.actionIdentifier else {
-            return
-        }
-        
-        start(cid: cid)
     }
 }
 
@@ -112,6 +97,7 @@ private extension DemoAppCoordinator {
     }
     
     func makeChatVC(for user: DemoUserType, startOn cid: ChannelId?, onLogout: @escaping () -> Void) -> UIViewController {
+        // Construct channel list query
         let channelListQuery: ChannelListQuery
         switch user {
         case let .credentials(userCredentials):
@@ -119,28 +105,25 @@ private extension DemoAppCoordinator {
         case .anonymous, .guest:
             channelListQuery = .init(filter: .equal(.type, to: .messaging))
         }
-        
-        let channelController = cid.map { ChatClient.shared.channelController(for: $0) }
-        let channelVC = channelController.map { makeChannelVC(controller: $0) }
-        let channelNVC = channelVC.map { UINavigationController(rootViewController: $0) }
-        
-        let selectedChannel = channelController?.channel
-        let channelListController = ChatClient.shared.channelListController(query: channelListQuery)
+
+        let tuple = makeChannelVCs(for: cid)
+        let selectedChannel = tuple.channelController?.channel
+        let channelListController = chat.channelListController(query: channelListQuery)
         let channelListVC = makeChannelListVC(
             controller: channelListController,
             selectedChannel: selectedChannel,
             onLogout: onLogout
         )
-        
+
         let channelListNVC = UINavigationController(rootViewController: channelListVC)
         let isIpad = UIDevice.current.userInterfaceIdiom == .pad
         if isIpad {
             let splitVC = UISplitViewController()
             splitVC.preferredDisplayMode = .oneBesideSecondary
-            splitVC.viewControllers = [channelListNVC, channelNVC].compactMap { $0 }
+            splitVC.viewControllers = [channelListNVC, tuple.channelNVC].compactMap { $0 }
             return splitVC
         } else {
-            channelVC.map { channelListNVC.pushViewController($0, animated: false) }
+            tuple.channelVC.map { channelListNVC.pushViewController($0, animated: false) }
             return channelListNVC
         }
     }
@@ -161,363 +144,46 @@ private extension DemoAppCoordinator {
         channelVC.channelController = controller
         return channelVC
     }
+
+    // Creates channel controller, channel VC and navigation controller for given channel id
+    private func makeChannelVCs(for cid: ChannelId?)
+        -> (channelController: ChatChannelController?, channelVC: UIViewController?, channelNVC: UINavigationController?) {
+        guard let cid = cid else {
+            return (nil, nil, nil)
+        }
+        // Get channel controller (model)
+        let controller = chat.channelController(for: cid)
+
+        // Create channel VC with given controller
+        let channelVC = controller.map { makeChannelVC(controller: $0) }
+        let channelNVC = channelVC.map { UINavigationController(rootViewController: $0) }
+
+        return (controller, channelVC, channelNVC)
+    }
 }
 
-// MARK: - Auth
+// MARK: - User Auth
 
 private extension DemoAppCoordinator {
-    func setUpChat() {
-        guard ChatClient.shared == nil else { return }
-        
-        // Set the log level
-        LogConfig.level = .warning
-        LogConfig.formatters = [
-            PrefixLogFormatter(prefixes: [.info: "ℹ️", .debug: "🛠", .warning: "⚠️", .error: "🚨"])
-        ]
-
-        // HTTP and WebSocket Proxy with Proxyman.app
-        if isStreamInternalConfiguration || AppConfig.shared.demoAppConfig.isAtlantisEnabled {
-            Atlantis.start()
-        } else {
-            Atlantis.stop()
-        }
-        
-        // Create Client
-        ChatClient.shared = ChatClient(config: AppConfig.shared.chatClientConfig)
-        
-        // Config
-        Components.default.channelListRouter = DemoChatChannelListRouter.self
-        Components.default.channelVC = DemoChatChannelVC.self
-        Components.default.messageContentView = DemoChatMessageContentView.self
-        Components.default.messageListDateSeparatorEnabled = true
-        Components.default.messageListDateOverlayEnabled = true
-        Components.default._messageListDiffingEnabled = isStreamInternalConfiguration
-        Components.default.messageActionsVC = DemoChatMessageActionsVC.self
-        Components.default.reactionsSorting = { $0.type.position < $1.type.position }
-
-        StreamRuntimeCheck.assertionsEnabled = isStreamInternalConfiguration
-        StreamRuntimeCheck._isLazyMappingEnabled = !isStreamInternalConfiguration
-
-        // Performance tracker
-        if isStreamInternalConfiguration {
-            PerformanceMonitor.shared().performanceViewConfigurator.options = [.performance]
-            PerformanceMonitor.shared().start()
-        }
-
-        let localizationProvider = Appearance.default.localizationProvider
-        Appearance.default.localizationProvider = { key, table in
-            let localizedString = localizationProvider(key, table)
-            
-            return localizedString == key
-                ? Bundle.main.localizedString(forKey: key, value: nil, table: table)
-                : localizedString
-        }
-    }
-    
-    func connect(user: DemoUserType, completion: @escaping (Error?) -> Void) {
-        switch user {
-        case let .credentials(userCredentials):
-            ChatClient.shared.connectUser(
-                userInfo: userCredentials.userInfo,
-                token: userCredentials.token,
-                completion: completion
-            )
-        case let .guest(userId):
-            ChatClient.shared.connectGuestUser(userInfo: .init(id: userId), completion: completion)
-        case .anonymous:
-            ChatClient.shared.connectAnonymousUser(completion: completion)
-        }
-    }
-    
     func logIn(as user: DemoUserType) {
+        // Store current user id
         UserDefaults.shared.currentUserId = user.staticUserId
-        
-        setUpChat()
-        
-        connect(user: user) { [weak self] in
-            if let error = $0 {
-                log.warning(error.localizedDescription)
-            } else {
-                self?.setupRemoteNotifications()
-            }
-        }
+
+        // App configuration used by our dev team
+        DemoAppConfiguration.setInternalConfiguration()
+
+        chat.logIn(as: user)
     }
     
     func logOut() {
-        let currentUserController = ChatClient.shared.currentUserController()
-        if let deviceId = currentUserController.currentUser?.currentDevice?.id {
-            currentUserController.removeDevice(id: deviceId) { error in
-                if let error = error {
-                    log.warning("removing the device failed with an error \(error)")
-                }
-            }
-        }
-        
-        ChatClient.shared.disconnect()
-        ChatClient.shared = nil
-        
+        // logout client
+        chat.logOut()
+
+        // clean user id
         UserDefaults.shared.currentUserId = nil
-        
+
+        // show login screen
         showLogin(animated: true)
-    }
-    
-    func setupRemoteNotifications() {
-        UNUserNotificationCenter
-            .current()
-            .requestAuthorization(options: [.alert, .sound, .badge]) { granted, _ in
-                guard granted else { return }
-                
-                DispatchQueue.main.async {
-                    UIApplication.shared.registerForRemoteNotifications()
-                }
-            }
-    }
-}
-
-// MARK: Custom Components for the Demo App
-
-final class DemoChatChannelVC: ChatChannelVC {
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    
-        let debugButton = UIBarButtonItem(
-            image: UIImage(systemName: "ladybug.fill")!,
-            style: .plain,
-            target: self,
-            action: #selector(debugTap)
-        )
-        navigationItem.rightBarButtonItems?.append(debugButton)
-    }
-    
-    @objc private func debugTap() {
-        guard let cid = channelController.cid else { return }
-        
-        let channelListVC: DemoChatChannelListVC
-        if let mainVC = splitViewController?.viewControllers.first as? UINavigationController,
-           let _channelListVC = mainVC.viewControllers.first as? DemoChatChannelListVC {
-            channelListVC = _channelListVC
-        } else if let _channelListVC = navigationController?.viewControllers.first as? DemoChatChannelListVC {
-            channelListVC = _channelListVC
-        } else {
-            return
-        }
-        
-        channelListVC.demoRouter.didTapMoreButton(for: cid)
-    }
-}
-
-final class DemoChatChannelListVC: ChatChannelListVC, EventsControllerDelegate {
-    /// The `UIButton` instance used for navigating to new channel screen creation.
-    lazy var createChannelButton: UIButton = {
-        let button = UIButton()
-        button.setImage(UIImage(systemName: "plus.message")!, for: .normal)
-        return button
-    }()
-
-    lazy var hiddenChannelsButton: UIButton = {
-        let button = UIButton()
-        button.setImage(UIImage(systemName: "archivebox")!, for: .normal)
-        return button
-    }()
-
-    private lazy var eventsController = controller.client.eventsController()
-    private lazy var connectionController = controller.client.connectionController()
-    private lazy var connectionDelegate = BannerShowingConnectionDelegate(
-        showUnder: navigationController!.navigationBar
-    )
-    
-    var demoRouter: DemoChatChannelListRouter {
-        router as! DemoChatChannelListRouter
-    }
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        eventsController.delegate = self
-        connectionController.delegate = connectionDelegate
-
-        navigationItem.rightBarButtonItems = [
-            UIBarButtonItem(customView: hiddenChannelsButton),
-            UIBarButtonItem(customView: createChannelButton)
-        ]
-        createChannelButton.addTarget(self, action: #selector(didTapCreateNewChannel), for: .touchUpInside)
-        hiddenChannelsButton.addTarget(self, action: #selector(didTapHiddenChannelsButton), for: .touchUpInside)
-    }
-
-    @objc private func didTapCreateNewChannel(_ sender: Any) {
-        demoRouter.showCreateNewChannelFlow()
-    }
-
-    @objc private func didTapHiddenChannelsButton(_ sender: Any) {
-        demoRouter.showHiddenChannels()
-    }
-    
-    var highlightSelectedChannel: Bool { splitViewController?.isCollapsed == false }
-    var selectedChannel: ChatChannel?
-
-    override func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-        let channel = controller.channels[indexPath.row]
-        selectedChannel = controller.channels[indexPath.row]
-        router.showChannel(for: channel.cid)
-    }
-
-    override func controller(_ controller: DataController, didChangeState state: DataController.State) {
-        super.controller(controller, didChangeState: state)
-
-        if highlightSelectedChannel && (state == .remoteDataFetched || state == .localDataFetched) && selectedChannel == nil {
-            guard let channel = self.controller.channels.first else { return }
-            
-            router.showChannel(for: channel.cid)
-            
-            selectedChannel = channel
-        }
-    }
-
-    override func controller(_ controller: ChatChannelListController, didChangeChannels changes: [ListChange<ChatChannel>]) {
-        super.controller(controller, didChangeChannels: changes)
-
-        guard highlightSelectedChannel else { return }
-        guard let selectedChannel = selectedChannel else { return }
-        guard let selectedChannelRow = controller.channels.firstIndex(of: selectedChannel) else {
-            return
-        }
-
-        let selectedItemIndexPath = IndexPath(row: selectedChannelRow, section: 0)
-
-        collectionView.selectItem(
-            at: selectedItemIndexPath,
-            animated: false,
-            scrollPosition: .centeredHorizontally
-        )
-    }
-
-    func eventsController(_ controller: EventsController, didReceiveEvent event: Event) {
-        if let newMessageEvent = event as? MessageNewEvent {
-            // This is a DemoApp integration test to make sure there are no deadlocks when
-            // accessing CoreDataLazy properties from the EventsController.delegate
-            _ = newMessageEvent.message.author
-        }
-    }
-}
-
-final class HiddenChannelListVC: ChatChannelListVC {
-    override func setUpAppearance() {
-        super.setUpAppearance()
-
-        title = "Hidden Channels"
-        navigationItem.leftBarButtonItem = nil
-    }
-}
-
-final class DemoChatMessageContentView: ChatMessageContentView {
-    override func updateContent() {
-        super.updateContent()
-
-        if content?.isShadowed == true {
-            textView?.textColor = appearance.colorPalette.textLowEmphasis
-            textView?.text = "This message is from a shadow banned user"
-        }
-
-        if let translations = content?.translations, let turkishTranslation = translations[.turkish] {
-            textView?.text = turkishTranslation
-            if let timestampLabelText = timestampLabel?.text {
-                timestampLabel?.text = "\(timestampLabelText) - Translated to Turkish"
-            }
-        }
-
-        guard let authorNameLabel = authorNameLabel, authorNameLabel.text?.isEmpty == true else {
-            return
-        }
-
-        guard let birthLand = content?.author.birthLand else {
-            return
-        }
-
-        authorNameLabel.text?.append(" \(birthLand)")
-    }
-}
-
-final class DemoChatMessageActionsVC: ChatMessageActionsVC {
-    // For the propose of the demo app, we add an extra hard delete message to test it.
-    override var messageActions: [ChatMessageActionItem] {
-        var actions = super.messageActions
-        if message?.isSentByCurrentUser == true && AppConfig.shared.demoAppConfig.isHardDeleteEnabled {
-            actions.append(hardDeleteActionItem())
-        }
-        actions.append(translateActionItem())
-        return actions
-    }
-
-    func hardDeleteActionItem() -> ChatMessageActionItem {
-        HardDeleteActionItem(
-            action: { [weak self] _ in
-                guard let self = self else { return }
-                self.alertsRouter.showMessageDeletionConfirmationAlert { confirmed in
-                    guard confirmed else { return }
-
-                    self.messageController.deleteMessage(hard: true) { _ in
-                        self.delegate?.chatMessageActionsVCDidFinish(self)
-                    }
-                }
-            },
-            appearance: appearance
-        )
-    }
-    
-    func translateActionItem() -> ChatMessageActionItem {
-        TranslateActionitem(
-            action: { [weak self] _ in
-                guard let self = self else { return }
-                self.messageController.translate(to: .turkish) { _ in
-                    self.delegate?.chatMessageActionsVCDidFinish(self)
-                }
-                
-            },
-            appearance: appearance
-        )
-    }
-
-    struct HardDeleteActionItem: ChatMessageActionItem {
-        var title: String { "Hard Delete Message" }
-        var isDestructive: Bool { true }
-        let icon: UIImage
-        let action: (ChatMessageActionItem) -> Void
-
-        init(
-            action: @escaping (ChatMessageActionItem) -> Void,
-            appearance: Appearance = .default
-        ) {
-            self.action = action
-            icon = appearance.images.messageActionDelete
-        }
-    }
-    
-    struct TranslateActionitem: ChatMessageActionItem {
-        var title: String { "Translate to Turkish" }
-        var isDestructive: Bool { false }
-        let icon: UIImage
-        let action: (ChatMessageActionItem) -> Void
-        
-        init(
-            action: @escaping (ChatMessageActionItem) -> Void,
-            appearance: Appearance = .default
-        ) {
-            self.action = action
-            icon = UIImage(systemName: "flag.fill")!
-        }
-    }
-}
-
-extension MessageReactionType {
-    var position: Int {
-        switch rawValue {
-        case "love": return 0
-        case "haha": return 1
-        case "like": return 2
-        case "sad": return 3
-        case "wow": return 4
-        default: return 5
-        }
     }
 }
