+++ DemoApp/Extensions/UIImage+Resized.swift
@@ -1,0 +1,29 @@
+//
+// Copyright © 2022 Stream.io Inc. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+extension UIImage {
+    func resized(to targetSize: CGSize) -> UIImage {
+        defer {
+            UIGraphicsEndImageContext()
+        }
+
+        let size = self.size
+        let widthRatio = targetSize.width / size.width
+        let heightRatio = targetSize.height / size.height
+        let newSize = widthRatio > heightRatio ? CGSize(width: size.width * heightRatio, height: size.height * heightRatio) :
+            CGSize(
+                width: size.width * widthRatio,
+                height: size.height * widthRatio
+            )
+        let rect = CGRect(x: 0, y: 0, width: newSize.width, height: newSize.height)
+
+        UIGraphicsBeginImageContextWithOptions(newSize, false, 1.0)
+        draw(in: rect)
+  
+        return UIGraphicsGetImageFromCurrentImageContext()!
+    }
+}
