+++ DemoApp/Screens/AppConfigViewController/SwitchButton.swift
