+++ DemoApp/Screens/Create Chat/Cells/GroupUserCell.swift
@@ -1,0 +1,12 @@
+//
+// Copyright © 2022 Stream.io Inc. All rights reserved.
+//
+
+import Foundation
+import StreamChatUI
+import UIKit
+
+class GroupUserCell: UICollectionViewCell {
+    @IBOutlet var nameLabel: UILabel!
+    @IBOutlet var avatarView: AvatarView!
+}
