+++ DemoApp/StreamChat/Components/Banner/BannerShowingConnectionDelegate.swift
