+++ DemoAppPush/NotificationService.swift
@@ -82,7 +82,7 @@ class NotificationService: UNNotificationServiceExtension {
         }
 
         var config = ChatClientConfig(apiKey: .init(apiKeyString))
-//        config.isLocalStorageEnabled = true
+//        config.isLocalStorageEnabled = false
         config.applicationGroupIdentifier = applicationGroupIdentifier
 
         let client = ChatClient(config: config)
